+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for refresh tokens in the Resource Owner Password Credentials flow and Authorization Code (with and without PKCE) flows.
+
+## [5.2.0] - 2020-10-14
+### Added
+- Allow to provide a `requests.Session` instance for `*AuthorizationCode` flows (even `PKCE`), `*ClientCredentials` and `*ResourceOwnerPasswordCredentials` flows.
+- Explicit support for Python 3.9
+
+### Changed
+- Code now follow `black==20.8b1` formatting instead of the git master version.
 
 ## [5.1.0] - 2020-03-04
 ### Added
@@ -122,7 +132,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Public release
 
-[Unreleased]: https://github.com/Colin-b/requests_auth/compare/v5.1.0...HEAD
+[Unreleased]: https://github.com/Colin-b/requests_auth/compare/v5.2.0...HEAD
+[5.2.0]: https://github.com/Colin-b/requests_auth/compare/v5.1.0...v5.2.0
 [5.1.0]: https://github.com/Colin-b/requests_auth/compare/v5.0.2...v5.1.0
 [5.0.2]: https://github.com/Colin-b/requests_auth/compare/v5.0.1...v5.0.2
 [5.0.1]: https://github.com/Colin-b/requests_auth/compare/v5.0.0...v5.0.1